Draggable marker added to the main map activity. Once the marker has stopped dragging the user has 10s to save the location, then the marker resets. Was too complex to load the map in another activity just for the sake of dragging the marker.
