--- app/src/main/res/layout/activity_adjust_location.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".AdjustLocationActivity">
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:background="?attr/colorPrimary"
-            android:layout_height="70dp"
-            app:title="Prototype" />
-
-        <!-- OSMDroid MapView for adjusting the marker -->
-        <org.osmdroid.views.MapView
-            android:id="@+id/adjustMap"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginTop="70dp"/>
-
-        <!-- Save button to confirm the adjusted location -->
-        <Button
-            android:id="@+id/confirm_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Save Adjusted Location"
-            android:layout_alignParentBottom="true"
-            android:layout_centerHorizontal="true"
-            android:padding="12dp"/>
-    </RelativeLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
